+++ packages/core/src/languages/index.ts
@@ -1,3 +1,8 @@
+export { default as BuiltInLanguagesPlugin } from "./builtin";
+export { default as CssLanguageSupport } from "./css";
+export { default as HtmlLanguageSupport } from "./html";
+export { default as JavaScriptLanguageSupport } from "./javascript";
+
 export type {
   ManglerExpression,
   ManglerExpressions,