+++ packages/core/src/manglers/__tests__/html-attributes.test.ts
@@ -31,6 +31,11 @@ suite("HTML Attribute Mangler", function() {
             input: ".data-foo { }",
             expected: ".data-foo { }",
           },
+          {
+            input: "[data-foo] { }",
+            expected: "[a] { }",
+            prefix: "",
+          },
         ],
       },
     ];
@@ -82,6 +87,11 @@ suite("HTML Attribute Mangler", function() {
             input: "<div class=\"data-foo\"></div>",
             expected: "<div class=\"data-foo\"></div>",
           }),
+          {
+            input: "<div data-foo=\"bar\"></div>",
+            expected: "<div a=\"bar\"></div>",
+            prefix: "",
+          },
         ],
       },
     ];
@@ -133,6 +143,11 @@ suite("HTML Attribute Mangler", function() {
             input: "document.querySelectorAll(\".data-foo\");",
             expected: "document.querySelectorAll(\".data-foo\");",
           }),
+          {
+            input: "document.querySelectorAll(\"[data-foo]\");",
+            expected: "document.querySelectorAll(\"[a]\");",
+            prefix: "",
+          },
         ],
       },
     ];