+++ packages/core/src/manglers/html-ids.ts
@@ -203,7 +203,11 @@ export default class HtmlIdMangler extends SimpleManglerPlugin {
   private static getPatterns(
     idNamePattern?: string | string[],
   ): string | string[] {
-    return idNamePattern || HtmlIdMangler.DEFAULT_PATTERNS;
+    if (idNamePattern === undefined) {
+      return HtmlIdMangler.DEFAULT_PATTERNS;
+    }
+
+    return idNamePattern;
   }
 
   /**
@@ -213,7 +217,11 @@ export default class HtmlIdMangler extends SimpleManglerPlugin {
    * @returns The reserved names to be used.
    */
   private static getReserved(reservedIds?: string[]): string[] {
-    return reservedIds || HtmlIdMangler.DEFAULT_RESERVED;
+    if (reservedIds === undefined) {
+      return HtmlIdMangler.DEFAULT_RESERVED;
+    }
+
+    return reservedIds;
   }
 
   /**
@@ -223,6 +231,10 @@ export default class HtmlIdMangler extends SimpleManglerPlugin {
    * @returns The prefix to be used.
    */
   private static getPrefix(keepIdPrefix?: string): string {
-    return keepIdPrefix || HtmlIdMangler.DEFAULT_PREFIX;
+    if (keepIdPrefix === undefined) {
+      return HtmlIdMangler.DEFAULT_PREFIX;
+    }
+
+    return keepIdPrefix;
   }
 }